- Added xcode project file for external glfw library + universal binaries (10.5 not 10.4) - orx embedded version now compiles using 10.5 base SDK (glfw needs HID manager defined in 10.5 version) - Switched default plugin list for embedded-mode to GLFW plugins

+++ libsndfile-1.0.22/include/sndfile.h
@@ -312,7 +312,7 @@ typedef	struct SNDFILE_tag	SNDFILE ;
 typedef __int64		sf_count_t ;
 //#define SF_COUNT_MAX		0x7fffffffffffffffi64
 #else
-typedef __int64	sf_count_t ;
+typedef long long	sf_count_t ;
 #define SF_COUNT_MAX		0x7FFFFFFFFFFFFFFFLL
 #endif
